Add files via upload
The analysis for the music store has been done using MySQL and Excel CSV files to understand the insights and retrieve the data. There are a total of 11 CSV format Excel files that have been downloaded from GitHub for analysis and further they have been uploaded to MySQL. The SQL queries are in a synchronized way with questions as the heading and the queries below the questions.